build: Updating the build process to use makefile
Based on commit 27b71778931c3f107eaaf5feacaaa5fa61d9c10f, the idea of
this commit is to transform the whole build chain to use makefile
instead of the actual hacky generate-dev-env.sh + custom scripts.

It have the benefits of improving parallelism, building in sub-directories and simple make commands.

This patch induce some moves in the directory scheme & the associated roles:
- "base" is now moved into the root (/) of the project as it's no more specific to a custom ceph release.
- "daemon-base" is now in charge of installing Ceph
- "daemon" is now in charge of specializing Ceph to the containerization
- distro version is integrated in the directory name like "jewel/centos/7.4.1708/"

The makefile have the default build targets that can be built in parallel.
This induce a build inside a temporary directory and a customization of the FROM directive.
The FROM is now defined to BASEIMAGE and substituted by the proper dependency at build time.

One of the benefits is also being able to define a custom repository at
build time like selecting this target :
    luminous@amd64@rhel@dummyrepository.redhat.com\\:8888@rhel7@7.3-released

<TBC>