makefile: Hard linking file instead of copying
When preparing the TEMP directory where the docker build will occur, the
actual code was doing a "cp -a" which means "--no-dereference -R --preserve=all".

This is an issue since we may have links in the source directory and
docker doesn't like that at all. It does want local files, not links.

This patch is about switch the "-a" to "-lR" which means hard linking file instead of preserving the links.

As a result, the "cp" time will be shorter as no file neede to be copied and symlinks disapears which makes "docker build" happy.